Tests/LibWeb/Text/expected/set-selection-inside-shadow-root.txt

Selection range count: 1
Selection start offset: 0, end offset: 1
